+++ src/common/channel-basex.ts
@@ -1,11 +1,11 @@
 // debug messages
 import { OutputChannel, window } from "vscode";
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const ver = require("@quodatum/xqlint").version;
+const verXQlint = require("@quodatum/xqlint").version;
 
-const _channel:OutputChannel = window.createOutputChannel("BaseX");
+const _channel:OutputChannel = window.createOutputChannel("BaseX",{log:true});
 
-function logdate(){
+export function logdate(){
     return (new Date()).toISOString().slice(0, 19).replace(/-/g, "/").replace("T", " ");
 }
 const replacerFunc = () => {
@@ -28,7 +28,7 @@ export function dump(obj :object) {
 
 export class channel {
     static log(msg: string) :void{
-        _channel.appendLine("["+logdate()+"] "+msg) 
+        _channel.appendLine(msg) 
     }
     static appendLine(msg: string) :void{
         _channel.appendLine(msg)
@@ -40,5 +40,5 @@ export class channel {
         _channel.show
     }
 }
-channel.log("started, XQLint version: "+ver);
+channel.log(`Activate vscode-basex(${ '??' }) *****************  XQLint(${ verXQlint })`);
 _channel.show