+++ src/constants.ts
@@ -1,13 +1,19 @@
 /* eslint-disable @typescript-eslint/no-namespace */
 export namespace commands {
-    export const evaluateXPath = "basexTools.evaluateXPath";
     export const executeXQuery = "basexTools.executeXQuery";
+    export const xqLintReport = "basexTools.xqLintReport";
+    export const clearDiagnostics ="basexTools.clearDiagnostics"
+
+    export const getCurrentXPath = "basexTools.getCurrentXPath";        
+    export const evaluateXPath = "basexTools.evaluateXPath";
+    // the original Altova targeted implementation
+    export const executeXQueryJosh = "basexTools.executeXQueryJosh";
+    
     export const formatAsXml = "basexTools.formatAsXml";
     export const xmlToText = "basexTools.xmlToText";
     export const textToXml = "basexTools.textToXml";
-    export const getCurrentXPath = "basexTools.getCurrentXPath";
+
     export const minifyXml = "basexTools.minifyXml";
-    export const xqLintReport = "basexTools.xqLintReport";
 }
 
 export namespace contextKeys {