+++ src/providers/documentlink.ts
@@ -10,7 +10,7 @@ import {
     ExtensionContext, languages, TextDocument, CancellationToken,
     DocumentLink, DocumentLinkProvider,Uri
 } from 'vscode';
-import { channel } from "../common/logger";
+import { channel } from "../common/channel-basex";
 import { XQLintFactory, importRange } from "../common/xqlint";
 import { languageIds } from "../constants";